requirements.txt
# Core Django
Django==4.2.27
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
django-cors-headers==4.4.0
dj-database-url==3.0.1
python-dotenv==1.0.1
Pillow==10.4.0
asgiref==3.8.1
sqlparse==0.5.5
PyJWT==2.9.0
requests==2.32.4
python-dateutil==2.9.0
tzdata==2025.3

# Google AI — let pip pick grpcio >= 1.75.1 required by Python 3.14
google-generativeai>=0.8.0
google-genai>=1.0.0

# Production
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
